requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
